Update README and add simple example: `SimpleViewController`.
