+++ Example/Theme/UIColor+Colors.swift
@@ -1,0 +1,20 @@
+//
+//  UIColor+Colors.swift
+//  Example
+//
+//  Created by Ben Shutt on 17/07/2020.
+//  Copyright © 2020 3 SIDED CUBE APP PRODUCTIONS LTD. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+// MARK: - UIColor + Colors
+
+extension UIColor {
+    
+    /// Override `.red` color
+    static let red = UIColor(
+        red: 231/255, green: 19/255, blue: 36/255, alpha: 1
+    )
+}
