+++ Example/Theme/UIImage+Images.swift
@@ -1,0 +1,22 @@
+//
+//  UIImage+Images.swift
+//  Example
+//
+//  Created by Ben Shutt on 17/07/2020.
+//  Copyright © 2020 3 SIDED CUBE APP PRODUCTIONS LTD. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UIImage {
+    
+    /// Circular information i
+    static let information = UIImage(named: "information-32")
+    
+    /// Simple black cross
+    static let cross = UIImage(named: "cross-32")
+    
+    /// Donations
+    static let donations = UIImage(named: "donations")
+}
