+++ README.md
@@ -1,55 +1,51 @@
+## PostManager
+
+Encaspsulates the logic of when and how `UIView`s should be posted and removed.
+Internally `PostManager` handles:
+1. Dismissal `Timer`s which, after a given time, request to remove a `UIView`
+2. A `PostGestureManager` for gesture related dismissals, e.g. pan.
+3. Serial `Queue` if required. 
+
+## UIViewPoster
+
+While `PostManager` handles the logic of when to post and remove a `UIView`, `UIViewPoster` handles the actual posting and subviews.
+Specifically, adding and removing the subview to the subview hierarchy, and animating when required.
+
+`PostManager` references a `UIViewPoster` communicating when to post and remove.
+There are a few common implementations of `UIViewPoster` provided in this framwork.
+
+## PostView
+
+A `UIView` for posting other `UIView`s in a serial manner.
+The animation of posted `UIView`s is via a `translation` from the top.
+
+Commonly, `PostView` is constrained to the top, leading, trailing edges of a `UIView` e.g. a `UIWindow`.
+When a subview is posted, the `PostView` constrains it's edges to the added subview.
+So, in this case, a posted `UIView` would define it's height either intrinsically or via explicit constraints, with its width determined by the `PostView`.
+
 ## MessageStackView
 
-A simple wrapper of a vertical `UIStackView` for posting and removing messages.  
-The most simple case is posting a `Message`, which in turn creates a `MessageView`, but custom `UIView`s are also supported.
+A simple wrapper of a vertical `UIStackView` for posting and removing `UIView`s.
+These `UIView`s, being a `UIStackView`, are part of the `MessageStackView`'s `arrangedSubviews`.
+The animation of posted `UIView`s is handled via the `UIStackView` by setting `isHidden`.
 
 ### Usage
 
 There is an Example project in the repository, but a simple use case would be:  
 
     class ViewController: UIViewController {
-    
-        /// `MessageManager` for posting `Message`s  
-        private let messageManager = MessageManager() 
         
-        /// Add `MessageStackView` to view hierarchy
-        override func viewDidLoad() {
-            super.viewDidLoad()
-            messageManager.addTo(.top(view))
+        override func viewDidAppear(_ animated: Bool) {
+            super.viewDidAppear(animated)
+            
+            postViewOrCreate().post(message: Message(
+                title: "Message Title",
+                subtitle: "This is message subtitle for detail",
+                leftImage: .information,
+                rightImage: nil
+            ))
         }
-
-        /// Event method  
-        private func onSomeEvent() {  
-            messageManager.post(message: Message(  
-                title: "This is a title",  
-                subtitle: "This is a subtitle"
-            ))  
-        }  
     }
-
-### MessageManager
-
-This class is responsible for posting `Message`s and `UIView`s on a `MessageStackView`.  
-The lifetime of these items is determined by the `MessageTime` specified on post.  
-By default the `MessageStackView` is not added to a view hierarchy. To do so, call `messageManager.addTo(MessageLayout)` or explicitly add the `MessageStackView` as a subview to a `UIView` and add corresponding constraints.  
-
-The `MessageManager` will look after all `Timer`s when a finite `MessageTime` is specified for dismissal of an item.
-
-### MessageTime
-
-`enum` to specify when a `UIView` is dismissed from the `MessageStackView`:  
-1. `.after(TimeInterval)` - dismiss after a given `TimeInterval` (seconds)  
-2. `.onTap` - dismiss when the view is tapped 
-3. `.never` - Once posted do not dismiss  
-
-### MessageLayout
-
-`enum` to specify common `MessageStackView` layouts:  
-1. `.top(UIView)` - Add the `MessageStackView` to the safe top, safe leading, and safe width of the given `UIView`  
-2. `.bottom(UIView)` - Add the `MessageStackView` to the safe bottom, safe leading, and safe width of the given `UIView`  
-
-A `MessageLayout` does not have to be used, they are merely common use cases. You may explicitly add the `MessageStackView` as a subview to another `UIView` and constrain if you want.  
-Call `messageManager.addTo(MessageLayout)` to add the `MessageStackView` to a `UIView` and constrain it w.r.t the `MessageLayout`.
 
 ### Message
 
@@ -59,18 +55,16 @@ Properties include:
 3. `leftImage` - Optional  
 3. `rightImage` - Optional  
 
-These can be posted on the `MessageManager`.
-A `Message` describes how a `MessageView` should look, which is the view which is posted on the `MessageStackView`.
+A `Message` describes how a `MessageView` should look, which is the view which is posted on the `UIViewPoster`.
 
-### Custom View
+### BadgeMessage
 
-The `MessageManager` supports custom `UIView`s.  
+`BadgeMessage` is similar to `Message` with an updated design and interface.
+
+### MessageStackView + Custom View
+
+The `MessageStackView` supports custom `UIView`s.  
 The `UIView`s width will be determined by the `UIStackView` as the `distribution` is `fill` and the `UIStackView` axis is `.vertical`.  
 However the height of the `UIView` is driven by the `UIView` itself, e.g. autolayout.  
   
 If, when posting and removing this `UIView` from the `MessageStackView` you want the animation to be smooth, then consider a "breakable" constraint for the height. Since the `UIStackView` will animate the `isHidden` property on the `UIView`, which will set an explict height during animation.
-
-### MessageConfiguration
-
-Property on the `MessageManager`, specify a custom look and feel of  `Message`s.  
-To determine how the `MessageConfiguration` applies to a view, conform to `MessageConfigurable`.
